tts-cpp: build_text_attention_cache — cast to int64_t before multiplying for ggml_reshape_2d
CodeQL cpp/integer-multiplication-cast-to-long flagged
`n_heads * head_dim` (both `int`, multiplied as `int` and then implicitly
converted to `int64_t` for `ggml_reshape_2d`'s shape argument). For
Supertonic's vector-estimator the values are 4 × 64 = 256 so there is
no actual overflow risk today, but a tts-cpp callsite that ever uses
larger n_heads / head_dim would silently truncate. Cast first to make
the multiplication 64-bit. No behaviour change for any current caller.

Alert was not introduced by this PR (line dates back to the original
tts-cpp add `ef840d5c3`) but surfaces on PR #31 because the surrounding
file was touched. Fixing here keeps the PR's CodeQL gate green.

+++ tts-cpp/src/supertonic_vector_estimator.cpp
@@ -1083,7 +1083,7 @@ void build_text_attention_cache(vector_text_attention_cache & cache,
 
     ggml_tensor * attn = ggml_flash_attn_ext(cache.ctx, q_in, k_in, v_in,
                                              nullptr, 1.0f/16.0f, 0.0f, 0.0f);
-    attn = ggml_reshape_2d(cache.ctx, attn, n_heads * head_dim, q_len);
+    attn = ggml_reshape_2d(cache.ctx, attn, static_cast<int64_t>(n_heads) * head_dim, q_len);
     ggml_tensor * ctx_tc = ggml_cont(cache.ctx, ggml_transpose(cache.ctx, attn));
     ggml_set_name(ctx_tc, "vector_attn_ctx"); ggml_set_output(ctx_tc);
     ggml_build_forward_expand(cache.gf, ctx_tc);
